Move the mock to a test file
